Added: TingleFragment (To replace  functionality of TingleActivity) Added: fragment_tingle.xml (To replace activity_tingle.xml) Refactored: TingleActivity to contain fragments Refactored: activity_tingle.xml so it contains a fragment instead Bug: Toast is broken Bug: Changing to listActivity is broken.
