+++ app/src/main/res/layout/activity_tingle.xml
@@ -1,122 +1,11 @@
 <?xml version="1.0" encoding="utf-8"?>
-<LinearLayout
+<FrameLayout
+    android:id="@+id/fragment_container"
     xmlns:android="http://schemas.android.com/apk/res/android"
     android:layout_height="match_parent"
     android:layout_width="match_parent"
     android:orientation="vertical"
     android:padding="@dimen/main">
 
-    <!-- Item display -->
-    <LinearLayout
-        android:layout_width="match_parent"
-        android:layout_height="match_parent"
-        android:orientation="vertical"
-        android:layout_weight="1.3">
-        <TextView
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:text="@string/headline"
-            android:textSize="18dp"
-            android:gravity="top"/>
 
-        <!-- bar -->
-        <View
-            android:layout_width="match_parent"
-            android:layout_height="2dp"
-            android:background="@color/colorLine">
-        </View>
-
-        <LinearLayout
-            android:layout_width="match_parent"
-            android:layout_height="match_parent"
-            android:orientation="vertical"
-            android:gravity="center">
-
-            <TextView
-                android:id="@+id/last_thing"
-                android:layout_width="wrap_content"
-                android:layout_height="wrap_content"
-                android:textSize="25dp"
-                android:layout_marginBottom="20dp"
-                android:gravity="center_horizontal"/>
-
-
-        </LinearLayout>
-
-
-
-
-    </LinearLayout>
-
-    <!-- bar -->
-    <View
-        android:layout_width="match_parent"
-        android:layout_height="2dp"
-        android:background="@color/colorLine">
-    </View>
-
-    <!-- Edit boxes and buttons -->
-    <LinearLayout
-        android:layout_width="match_parent"
-        android:layout_height="match_parent"
-        android:orientation="vertical"
-        android:gravity="bottom"
-        android:layout_weight="1">
-
-
-        <LinearLayout
-            android:layout_width="match_parent"
-            android:layout_height="wrap_content"
-            android:orientation="horizontal"
-            android:paddingTop="5dp">
-            <TextView
-                android:layout_marginTop="10dp"
-                android:layout_width="wrap_content"
-                android:layout_height="match_parent"
-                android:text="@string/what_thing"/>
-            <EditText
-                android:id="@+id/what_text"
-                android:layout_width="match_parent"
-                android:layout_height="wrap_content"/>
-        </LinearLayout>
-
-
-        <LinearLayout
-            android:layout_width="match_parent"
-            android:layout_height="wrap_content"
-            android:orientation="horizontal"
-            android:paddingBottom="5dp">
-            <TextView
-                android:layout_marginTop="10dp"
-                android:layout_width="wrap_content"
-                android:layout_height="match_parent"
-                android:text="@string/where"
-                android:paddingRight="27dp"/>
-            <EditText
-                android:id="@+id/where_text"
-                android:layout_width="match_parent"
-                android:layout_height="wrap_content"/>
-        </LinearLayout>
-
-
-        <Button
-            android:id="@+id/add_button"
-            android:layout_width="match_parent"
-            android:layout_height="wrap_content"
-            android:text="@string/heading_create"/>
-
-        <Button
-            android:id="@+id/lookUp_button"
-            android:layout_width="match_parent"
-            android:layout_height="wrap_content"
-            android:text="@string/heading_lookUp"/>
-
-        <Button
-            android:id="@+id/goTOList_button"
-            android:layout_width="match_parent"
-            android:layout_height="wrap_content"
-            android:text="@string/heading_showAll"/>
-
-    </LinearLayout>
-
-</LinearLayout>
+</FrameLayout>
